Combine baggage packages from checkhttp and envoy
In this PR, we remove `Inject` method from `Propagator` interface as different propagators return different things from `Inject`. Instead, in `Check` call of both packages, we check if their propagators implement the `Inject` method and then call it.
