+++ pkg/policies/flowcontrol/baggage/baggage_suite_test.go
@@ -1,4 +1,4 @@
-package baggage_test
+package baggage
 
 import (
 	"testing"
@@ -9,7 +9,6 @@ import (
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
 	flowlabel "github.com/fluxninja/aperture/pkg/policies/flowcontrol/label"
-	"github.com/fluxninja/aperture/pkg/policies/flowcontrol/service/envoy/baggage"
 	"github.com/fluxninja/aperture/pkg/utils"
 )
 
@@ -37,7 +36,7 @@ var _ = AfterSuite(func() {
 })
 
 var _ = Describe("Prefixed propagator", func() {
-	var propagator baggage.Propagator = baggage.Prefixed{Prefix: "myprefix-"}
+	propagator := Prefixed{Prefix: "myprefix-"}
 
 	It("extracts flow labels from baggage", func() {
 		Expect(propagator.Extract(map[string]string{
@@ -77,7 +76,7 @@ var _ = Describe("Prefixed propagator", func() {
 })
 
 var _ = Describe("W3 Baggage propagator", func() {
-	var propagator baggage.Propagator = baggage.W3Baggage{}
+	propagator := W3Baggage{}
 
 	Context("when there's no baggage header", func() {
 		It("reads no flow labels", func() {
@@ -112,7 +111,7 @@ var _ = Describe("W3 Baggage propagator", func() {
 		It("creates injecting instructions for envoy", func() {
 			newHeaders, err := propagator.Inject(flowlabel.FlowLabels{
 				"hello": fl("world"),
-			}, baggage.Headers(map[string]string{
+			}, Headers(map[string]string{
 				"baggage":      "foo=bar,baz=quux",
 				"content-type": "application/json",
 			}))
@@ -149,6 +148,91 @@ var _ = Describe("W3 Baggage propagator", func() {
 				Header: &envoy_core.HeaderValue{Key: "baggage", Value: "foo=bar"},
 				Append: wrapperspb.Bool(false),
 			}}))
+		})
+	})
+
+	It("ignores member properties", func() {
+		Expect(propagator.Extract(map[string]string{
+			"baggage": "foo=bar;props",
+		})).To(Equal(flowlabel.FlowLabels{
+			"foo": fl("bar"),
+		}))
+	})
+
+	It("handles urlencoded values", func() {
+		Expect(propagator.Extract(map[string]string{
+			// TODO make sure this is correct FLUX-1290
+			"baggage": "foo=%2520",
+		})).To(Equal(flowlabel.FlowLabels{
+			"foo": fl(" "),
+		}))
+	})
+})
+
+var _ = Describe("W3 Baggage propagator", func() {
+	propagator := W3BaggageCheckHTTP{}
+
+	Context("when there's no baggage header", func() {
+		It("reads no flow labels", func() {
+			Expect(propagator.Extract(map[string]string{
+				"content-type": "application/json",
+			})).To(Equal(flowlabel.FlowLabels{}))
+		})
+
+		It("creates injecting instructions for envoy", func() {
+			newHeaders, err := propagator.Inject(flowlabel.FlowLabels{
+				"foo": fl("bar"),
+			}, nil)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(newHeaders).To(Equal(map[string]string{"baggage": "foo=bar"}))
+		})
+	})
+
+	Context("when baggage header exists", func() {
+		It("extracts flow labels from baggage", func() {
+			Expect(propagator.Extract(map[string]string{
+				"baggage":      "foo=bar,baz=quux",
+				"content-type": "application/json",
+			})).To(Equal(flowlabel.FlowLabels{
+				"foo": fl("bar"),
+				"baz": fl("quux"),
+			}))
+		})
+
+		It("creates injecting instructions for envoy", func() {
+			newHeaders, err := propagator.Inject(flowlabel.FlowLabels{
+				"hello": fl("world"),
+			}, Headers(map[string]string{
+				"baggage":      "foo=bar,baz=quux",
+				"content-type": "application/json",
+			}))
+			Expect(err).NotTo(HaveOccurred())
+			Expect(newHeaders).To(Equal(map[string]string{"baggage": "foo=bar,baz=quux,hello=world"}))
+		})
+	})
+
+	Context("when flow label has 'hidden' flag", func() {
+		It("extracts it correctly", func() {
+			Expect(propagator.Extract(map[string]string{
+				"baggage":      "foo=bar",
+				"content-type": "application/json",
+			})).To(Equal(flowlabel.FlowLabels{
+				"foo": flowlabel.FlowLabelValue{
+					Value:     "bar",
+					Telemetry: true,
+				},
+			}))
+		})
+
+		It("injects it correctly", func() {
+			newHeaders, err := propagator.Inject(flowlabel.FlowLabels{
+				"foo": flowlabel.FlowLabelValue{
+					Value:     "bar",
+					Telemetry: true,
+				},
+			}, nil)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(newHeaders).To(Equal(map[string]string{"baggage": "foo=bar"}))
 		})
 	})
